control-plane/migration-coordinator/requirements.txt
kubernetes~=18.20.0
Flask~=2.0.2
click~=8.1.3
requests~=2.27.1
python-dateutil~=2.8.2
PyYAML~=6.0